Do not persist notifications
Persistent notifications cause confusing messages shown to the user
about the mail not being up-to-date due to storing the notification state
to the server-side. This is not necessarily often the case for regular
users since it happens only if one uses multiple different Outlook
clients at the same time (like Outlook Web Access and the desktop app).